Oops forgot to turn disabled keys off.
It won't effect server.twsicommunity.com because channels will be locked knowing that users wouldn't be able to do much. With that said it's probably useless to use inspect elements thing. Not that people will use other's stream key just for that.... As said before though, Remember to lock your streams before even going off it.